# account_manager/openapi/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: vasco.m.campos@inesctec.pt
    name: Vasco Campos
  description: 'User Account Manager Service OpenAPI definition. This service has
    the following functions: register, login, change user credentials and update it''s
    profile settings. Find out more: [User Account Manager Service documentation](https://gitlab.inesctec.pt/cpes/european-projects/interconnect/hems/hems-documentation/-/blob/master/Microservices/User-Account-Manager-Service.adoc)'
  title: User Account Manager Service
  version: 1.2.0
servers:
- url: https://interconnect-dev.inesctec.pt/api/account
security:
- {}
- Bearer: []
...
paths:
...
  /list-dongles:
    description: Retrieve information about dongles api keys
    get:
      operationId: list_dongles_get
      parameters:
      - explode: false
        in: header
        name: X-Correlation-ID
        required: true
        schema:
          $ref: '#/components/schemas/CorrelationId'
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Dongles'
                maxItems: 1000
                type: array
          description: List of installed dongles
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "400":
          content:
            application/json:
              examples:
                Bad format:
                  value:
                    error: Invalid specification format
                Timestamp error:
                  value:
                    error: one or more timestamps in the request are invalid
                Missing X-Correlation-ID header:
                  value:
                    error: missing X-Correlation-ID header
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "403":
          content:
            application/json:
              examples:
                Forbidden:
                  value:
                    error: you are not authorized to access this information
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden to access the information requested
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
      summary: Get information about dongles from users
      tags:
      - User info management
      x-openapi-router-controller: account_manager.controllers.user_info_management_controller
  /list-meter-ids:
    description: Retrieve information about meter ids
    get:
      operationId: list_meter_ids_get
      parameters:
      - explode: false
        in: header
        name: X-Correlation-ID
        required: true
        schema:
          $ref: '#/components/schemas/CorrelationId'
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/MeterId'
                maxItems: 1000
                type: array
          description: List of installed meter ids
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "400":
          content:
            application/json:
              examples:
                Bad format:
                  value:
                    error: Invalid specification format
                Timestamp error:
                  value:
                    error: one or more timestamps in the request are invalid
                Missing X-Correlation-ID header:
                  value:
                    error: missing X-Correlation-ID header
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "403":
          content:
            application/json:
              examples:
                Forbidden:
                  value:
                    error: you are not authorized to access this information
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden to access the information requested
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
      summary: Get information about meter ids from users
      tags:
      - User info management
      x-openapi-router-controller: account_manager.controllers.user_info_management_controller
...
  /meter-to-user:
    description: Retrieve the user ids that contains the specified meter ids
    get:
      operationId: meter_to_user_get
      parameters:
      - explode: false
        in: header
        name: X-Correlation-ID
        required: true
        schema:
          $ref: '#/components/schemas/CorrelationId'
        style: simple
      - explode: false
        in: header
        name: Authorization
        required: false
        schema:
          $ref: '#/components/schemas/Authorization'
        style: simple
      - explode: true
        in: query
        name: meter-ids
        required: true
        schema:
          example:
          - meterId1
          - meterId2
          items:
            $ref: '#/components/schemas/MeterId'
          minItems: 1
          type: array
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/MeterIdToUserId'
                minItems: 1
                type: array
          description: user ids that corresponds to requested meter ids
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "400":
          content:
            application/json:
              examples:
                Bad format:
                  value:
                    error: Invalid specification format
                Timestamp error:
                  value:
                    error: one or more timestamps in the request are invalid
                Missing X-Correlation-ID header:
                  value:
                    error: missing X-Correlation-ID header
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "401":
          content:
            application/json:
              examples:
                Unauthenticated:
                  value:
                    error: error in authorization token
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthenticated or invalid credentials
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "403":
          content:
            application/json:
              examples:
                Forbidden:
                  value:
                    error: you are not authorized to access this information
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden to access the information requested
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "404":
          content:
            application/json:
              examples:
                User Id not found:
                  value:
                    error: user with id "userId" not found
              schema:
                $ref: '#/components/schemas/Error'
          description: User id not found
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
      summary: Retrieve the user ids that contains the specified meter ids
      tags:
      - User info management
      x-openapi-router-controller: account_manager.controllers.user_info_management_controller
...
  /user-deleted-list:
    description: Retrieve list of all deleted users
    get:
      operationId: user_deleted_list_get
      parameters:
      - explode: false
        in: header
        name: X-Correlation-ID
        required: true
        schema:
          $ref: '#/components/schemas/CorrelationId'
        style: simple
      - explode: false
        in: header
        name: Authorization
        required: false
        schema:
          $ref: '#/components/schemas/Authorization'
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/DeletedUser'
                type: array
          description: All deleted users
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "400":
          content:
            application/json:
              examples:
                Bad format:
                  value:
                    error: Invalid specification format
                Timestamp error:
                  value:
                    error: one or more timestamps in the request are invalid
                Missing X-Correlation-ID header:
                  value:
                    error: missing X-Correlation-ID header
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "401":
          content:
            application/json:
              examples:
                Unauthenticated:
                  value:
                    error: error in authorization token
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthenticated or invalid credentials
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "403":
          content:
            application/json:
              examples:
                Forbidden:
                  value:
                    error: you are not authorized to access this information
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden to access the information requested
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
      summary: Retrieve list of all deleted users
      tags:
      - User info management
      x-openapi-router-controller: account_manager.controllers.user_info_management_controller
  /user-list:
    description: Retrieve list of all users ids
    get:
      operationId: user_list_get
      parameters:
      - explode: false
        in: header
        name: X-Correlation-ID
        required: true
        schema:
          $ref: '#/components/schemas/CorrelationId'
        style: simple
      - explode: false
        in: header
        name: Authorization
        required: false
        schema:
          $ref: '#/components/schemas/Authorization'
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/UserId'
                type: array
          description: All user ids
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "400":
          content:
            application/json:
              examples:
                Bad format:
                  value:
                    error: Invalid specification format
                Timestamp error:
                  value:
                    error: one or more timestamps in the request are invalid
                Missing X-Correlation-ID header:
                  value:
                    error: missing X-Correlation-ID header
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "401":
          content:
            application/json:
              examples:
                Unauthenticated:
                  value:
                    error: error in authorization token
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthenticated or invalid credentials
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "403":
          content:
            application/json:
              examples:
                Forbidden:
                  value:
                    error: you are not authorized to access this information
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden to access the information requested
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
      summary: Retrieve list of all users ids
      tags:
      - User info management
      x-openapi-router-controller: account_manager.controllers.user_info_management_controller
components:
  headers:
    CorrelationId:
      explode: false
      required: true
      schema:
        $ref: '#/components/schemas/CorrelationId'
      style: simple
...
  schemas:
    CorrelationId:
      format: uuid
      type: string
    UserId:
      example: userId1
      maxLength: 10
      minLength: 10
      title: user_id
      type: string
...
    Authorization:
      format: byte
      type: string
...
    MeterId:
      title: MeterId
      type: string
    Country:
      enum:
      - PT
      example: PT
      title: Country
      type: string
...
    ScheduleType:
      enum:
      - economic
      - ecologic
      example: economic
      title: ScheduleType
      type: string
    TarifType:
      enum:
      - simple
      - bi-hourly
      - tri-hourly
      example: simple
      title: TarifType
      type: string
    ContractedPower:
      enum:
      - 1.15 kVA
      - 2.3 kVA
      - 3.45 kVA
      - 4.6 kVA
      - 5.75 kVA
      - 6.9 kVA
      - 10.35 kVA
      - 13.8 kVA
      - 17.25 kVA
      - 20.7 kVA
      - 27.6 kVA
      - 34.5 kVA
      - 41.4 kVA
      - 27.6 kVA
      - 34.5 kVA
      - 41.4 kVA
      example: 6.9 kVA
      title: ContractedPower
      type: string
...
    Permission:
      enum:
      - None
      - Minimal
      - Full
      example: Full
      title: Permission
      type: string
...
    DeletedUser:
      description: Settings object
      example:
        country: PT
        is_active: true
        schedule_type: economic
        not_disturb:
          sunday:
          - end_timestamp: 2000-01-23T04:56:07.000+00:00
            start_timestamp: 2000-01-23T04:56:07.000+00:00
          - end_timestamp: 2000-01-23T04:56:07.000+00:00
            start_timestamp: 2000-01-23T04:56:07.000+00:00
          saturday:
          - end_timestamp: 2000-01-23T04:56:07.000+00:00
            start_timestamp: 2000-01-23T04:56:07.000+00:00
          - end_timestamp: 2000-01-23T04:56:07.000+00:00
            start_timestamp: 2000-01-23T04:56:07.000+00:00
          tuesday:
          - end_timestamp: 2000-01-23T04:56:07.000+00:00
            start_timestamp: 2000-01-23T04:56:07.000+00:00
          - end_timestamp: 2000-01-23T04:56:07.000+00:00
            start_timestamp: 2000-01-23T04:56:07.000+00:00
          wednesday:
          - end_timestamp: 2000-01-23T04:56:07.000+00:00
            start_timestamp: 2000-01-23T04:56:07.000+00:00
          - end_timestamp: 2000-01-23T04:56:07.000+00:00
            start_timestamp: 2000-01-23T04:56:07.000+00:00
          thursday:
          - end_timestamp: 2000-01-23T04:56:07.000+00:00
            start_timestamp: 2000-01-23T04:56:07.000+00:00
          - end_timestamp: 2000-01-23T04:56:07.000+00:00
            start_timestamp: 2000-01-23T04:56:07.000+00:00
          friday:
          - end_timestamp: 2000-01-23T04:56:07.000+00:00
            start_timestamp: 2000-01-23T04:56:07.000+00:00
          - end_timestamp: 2000-01-23T04:56:07.000+00:00
            start_timestamp: 2000-01-23T04:56:07.000+00:00
          monday:
          - end_timestamp: 2000-01-23T04:56:07.000+00:00
            start_timestamp: 2000-01-23T04:56:07.000+00:00
          - end_timestamp: 2000-01-23T04:56:07.000+00:00
            start_timestamp: 2000-01-23T04:56:07.000+00:00
        is_google_account: true
        last_name: Ana
        wp_token: token
        modified_timestamp: 2000-01-23T04:56:07.000+00:00
        contracted_power: 6.9 kVA
        deleted: true
        expo_token: I22V7MDFJx
        user_id: userId1
        api_key: api_key
        permissions: Full
        global_optimizer: true
        postal_code: postal_code
        first_name: Ana
        meter_id: meter_id
        deleted_timestamp: 2000-01-23T04:56:07.000+00:00
        email: myemail@mycompany.com
        tarif_type: simple
      properties:
        user_id:
          example: userId1
          maxLength: 10
          minLength: 10
          title: user_id
          type: string
        first_name:
          example: Ana
          maxLength: 15
          minLength: 1
          title: first_name
          type: string
        last_name:
          example: Ana
          maxLength: 15
          minLength: 1
          title: first_name
          type: string
        email:
          example: myemail@mycompany.com
          format: email
          maxLength: 50
          title: email
          type: string
        is_active:
          title: is_active
          type: boolean
        schedule_type:
          $ref: '#/components/schemas/ScheduleType'
        meter_id:
          title: MeterId
          type: string
        api_key:
          title: api_key
          type: string
        expo_token:
          example: I22V7MDFJx
          title: expo_token
          type: string
        wp_token:
          example: token
          title: wp_token
          type: string
        country:
          $ref: '#/components/schemas/Country'
        postal_code:
          pattern: ^[0-9]{4}-[0-9]{3}$
          title: postal_code
          type: string
        tarif_type:
          $ref: '#/components/schemas/TarifType'
        contracted_power:
          $ref: '#/components/schemas/ContractedPower'
        not_disturb:
          $ref: '#/components/schemas/NotDisturb'
        global_optimizer:
          title: global_optimizer
          type: boolean
        permissions:
          $ref: '#/components/schemas/Permission'
        is_google_account:
          example: true
          title: is_google_account
          type: boolean
        modified_timestamp:
          format: date-time
          title: start_timestamp
          type: string
        deleted:
          title: deleted
          type: boolean
        deleted_timestamp:
          format: date-time
          title: start_timestamp
          type: string
      required:
      - deleted
      - deleted_timestamp
      - email
      - first_name
      - is_active
      - is_google_account
      - modified_timestamp
      - not_disturb
      - permissions
      - user_id
      title: DeletedUser
      type: object
...
    MeterIdToUserId:
      example:
        user_id: userId1
        meter_id: meter_id
      properties:
        meter_id:
          title: MeterId
          type: string
        user_id:
          example: userId1
          maxLength: 10
          minLength: 10
          title: user_id
          type: string
      required:
      - meter_id
      - user_id
      title: MeterIdToUserId
      type: object
    Dongles:
      example:
        user_id: userId1
        api_key: api_key
      properties:
        user_id:
          example: userId1
          maxLength: 10
          minLength: 10
          title: user_id
          type: string
        api_key:
          title: api_key
          type: string
      title: Dongles
      type: object
...
    Error:
      description: Error model
      properties:
        error:
          title: error
          type: string
      required:
      - error
      title: Error
      type: object
  securitySchemes:
    Bearer:
      bearerFormat: JWT
      scheme: bearer
      type: http
      x-bearerInfoFunc: account_manager.controllers.security_controller_.info_from_Bearer